backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-frontmatter==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
gitpython==3.1.43
